WIP: respect the name_len
The name_len isn't just a convenience, the strings we receive from drgn
may be substrings. Thus, our comparison functions need to directly use
name_len (not name_len+1 which assumes nul-term), and then verify that
the BTF string is nul-terminated, in order to verify that the BTF string
doesn't simply start with the name we are comparing.