WIP: Implement preliminary CTF type finder
This commit is not at all ready for submission but it gets the basic
gist of what is necessary to support CTF via libctf. For now, we load a
CTF archive given in an environment variable, only if we've already
successfully loaded kallsyms symbols. So, you need to run this as:

    DRGN_CTF_FILE=foo python -m drgn -c bar

This commit implements only a type finder. Still missing is the "object
finder" which can actually take a symbol address, determine its type,
and create the drgn.Object representing it. This is currently feasible
for data/variables, but functions are not yet contained in vmlinux.ctfa
archives, at least in UEK. The next piece of work will be to implement
this object finder.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>