Implement basic CTF type finder
This is a quite rough implementation of a CTF type finder. It is not
quite production-ready. We expose a Python binding to allow loading a
CTF file.

This commit implements only a type finder. Still missing is the "object
finder" which can actually take a symbol address, determine its type,
and create the drgn.Object representing it. This is currently feasible
for data/variables, but functions are not yet contained in vmlinux.ctfa
archives, at least in UEK. The next piece of work will be to implement
this object finder.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>