libdrgn: Use Symbol Finder API in find_symbol_by_address_internal()
The drgn_program_find_symbol_by_address_internal() function is the
workhorse for all the cases where libdrgn itself may want to lookup a
symbol: in particular, when formatting stack traces or formatting
objects. It does less work by possibly already having a Dwfl_Module
looked up, and by avoiding memory allocation of a symbol, and it's more
convenient because it doesn't return an error. Another difference is
that it does not return an error on lookup failure, since this is
frequently a normal occurrence, so it would waste an allocation.

Unfortunately, the new Symbol finder breaks two of these properties: the
returned symbol is now allocated via malloc(), and errors can be
returned via the lookup API. Still, the lookup API does allow us to keep
the already-determined module, if it is known.

Switch the internal method over to the new finder API. The major
difference now is simply that lookup failures don't result in an error.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>