python: Allow construction of Symbol objects
Previously, Symbol objects could not be constructed in Python. However,
in order to allow Python Symbol finders, this needs to be changed. We
introduce the concept of "owned" names inside symbols: ideally, a symbol
name shares the lifetime of the program associated with it. However, if
that's not the case (as with Python symbols), then the symbol itself
owns the name, and must free it when the symbol is freed. When the
symbol, rather than the program, owns the name buffer, the name_owned
field is true.

The added complexity is justified by the fact that most symbols are from
the ELF file, and thus share a lifetime with the Program. It would be a
waste to constantly strdup() these strings, just to support a small
number of Symbols created by Python code.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>