ctf: Implement DRGN_FIND_OBJECT_CONSTANT
Prior to this, we ignored the find object flags, and simply searched the
"vmlinux" dict for any variable we could find. Now, we respect the
flags. We don't default to any particular dict (which allows lookup of
module variables, but they may override vmlinux). Most notably, we build
an index of enumerators, and we search this when we are given
DRGN_FIND_OBJECT_CONSTANT. This fixes a fair amount of drgn
functionality on CTF. The major missing piece, at this point, is stack
unwinding.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>