ctf: Fix lookup of int/float types by name
When looking up int/float types by name, there can be many matches,
since there may be several copies of the type which are used in
different bit fields, which have different offsets and bit sizes.
Unfortunately, CTF reports their byte sizes differently from the
standard integer type. If an int/float is looked up by name, and it
appears in a bitfield type, it's likely that we'll get the bitfield type
ID, rather than the base type ID. And the result is that we may get a
type with the wrong size: potentially very bad.

Fix this by identifying the "canonical atomic types" for each name at
initialization time. The criteria are: the type with offset 0, bit size
== CTF byte size, and largest byte size, breaking ties with the lowest
CTF type ID. Use this mapping whenever a name-lookup results in a
INTEGER or FLOAT type.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>