scripts: Build drgn wheels with libctf
This is not intended for upstream. Instead, this simply enables me to
build libctf-enabled drgn wheels for distribution.

libctf is part of binutils, but only suitably recent versions of
binutils have it (and there are some major bugfix changes in more recent
versions). So the latest is best, however many distribution packages
don't make it easy to link to libctf statically. The best way is to
simply build and install binutils, and then link to the static
libraries.

For some reason, libiberty.a is not installed, so we need to go the
extra mile of setting up a linker path.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>