kallsyms: Support long names for Rust symbols
Since kernel commit 73bbb94466fd3 ("kallsyms: support "big" kernel
symbols"), symbol names could span two bytes using an encoding not
unlike UTF-8. This introduces ambiguity in the parsing code: the best
way to resolve it is to use the osrelease value, since the commit is not
likely to get backported to stable releases.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>