Optimize module symbol loading
On some systems, /proc/kcore reads are so slow, that reading module
symbols can take a significant amount of time. The reason is that the
Python code is making significantly more, and smaller, reads than it
needs to make.

It's more efficient to read all the data we need at once, rather than a
hundred small reads. Read the symbol table array, and the string array,
in a single go, rather than reading the entries we need individually.
On the system where we observed the issue, this improved runtime from
around 7 minutes to 4 seconds.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>