Trim the CI Python versions and include 3.12 beta
With 3.12 beta, there are 7 Python versions to test, which is making the
test matrix too large. To reduce the size and frequency of the CI runs,
we will only run 3: the oldest, the newest, and the current
pre-release/beta, if applicable. We will also stop running CI on push
events to feature branches: only pull requests will be tested.

The simplest way to do this is to create a new "reusable workflow"
called vmtest. We can specify whether to run on all python versions as
an input. The CI workflow will trigger tests, as will a push event when
the main branch is pushed.

Manual runs can be triggered via the UI once this is merged into the
main branch.

Signed-off-by: Stephen Brennan <stephen@brennan.io>