kallsyms: don't use proc unless we are root
Non-root /proc/kallsyms reports all addresses as 0. We could try to
detect this without the geteuid(), by reading several lines of
/proc/kallsyms and checking whether there is a non-null address.
However, that adds a fair bit of code complexity and doesn't provide
much benefit. So just check whether we are root: if not, fall back to
the VMCOREINFO and parsing method.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Reviewed-by: Imran Khan <imran.f.khan@oracle.com>