orc_info: Add built-in ORC unwinding
The existing use of ORC within drgn has been limited to using it as a
back-up. The ORC data was loaded from the ELF file associated with each
module, and it was only used when DWARF CFI rules were not
present (excepting the environment variables used for testing). Loading
the ORC from the ELF debuginfo files means that ORC cannot be used when
ELF debuginfo is not present, which is possible when using CTF or BTF
and kallsyms.

Since the point of ORC is that it can be used by the kernel itself to
unwind stacks, it is stored in kernel memory, and we can load it from
the core dump itself. The challenge is that the ORC unwinding is quite
integrated into the drgn_module debuginfo system, and drgn_module is
only created when an ELF file is available for a module. This is likely
to change with the upcoming new module API, but for now, we need another
way to locate the ORC info for a given PC.

To do this, introduce a binary search tree, independent of the
drgn_module subsystem, which maps PC ranges to relevant ORC info. This
is roughly equivalent to what dwfl_addrmodule() does for locating a
drgn_module. In the stack unwinding code, we allow unwinding with CFI
rules to continue even if a drgn_module is not found, so long as we can
find an ORC mapping range to use.

Similar to kallsyms and CTF, we do not automatically load any of this
information. Instead, a helper is exposed which allows us to load ORC
info. When CTF is in use, we can load built-in ORC data. In the standard
case where DWARF debuginfo is in-use, the existing unwind mechanisms
remain unchanged.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>