libdrgn: symbol: return &drgn_not_found
Technically, any drgn_error with a code DRGN_ERROR_LOOKUP is a valid
lookup error. However, the common practice has been to return
&drgn_not_found, which has the benefit of being statically allocated: no
cleanup is required in case the error should be ignored.

Unfortunately, the drgn ELF symbol finders are not using
&drgn_not_found, which breaks this common practice. Switch them to the
common practice so that the upcoming commit can continue.

A future commit might change the common practice to allow a function
such as "drgn_error_catch()". However, that's outside the scope of this
work.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Reviewed-by: Imran Khan <imran.f.khan@oracle.com>