orc_info: decouple from drgn_module
The drgn_module_orc_info is embedded within drgn_module, and most
functions in orc_info simply accept a struct drgn_module *. However, in
anticipation of the ability to use ORC without a debuginfo module, we
need to be able to operate on ORC info which is not embedded in a
drgn_module. So, do some rearranging which allows some core ORC
functions to work without the drgn_module or ELF related objects.

No functional change is intended.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>