Implement CTF type and object finder for Linux
The Compact C Type Format (CTF) is an alternative debug information
format which describes the types of C data structures. Unlike other
debug information formats (like DWARF), it is designed to be very
compact, so that it can be included in production binaries and packages:
NOT in debuginfo packages.

Support for generating CTF exists in GCC, and support for consuming CTF
exists in GDB. The Linux kernel does not currently support building with
CTF, but an effort is underway to change this. Oracle UEK kernels
already build with CTF: each kernel RPM contains a file named
"vmlinux.ctfa" which is a CTF archive: it contains type information for
the kernel and all modules.

CTF can be used as an alternative to DWARF. It cannot provide all the
features that DWARF would - in particular, stack traces require frame
pointers, or ORC, or DWARF CFI. CTF cannot allow drgn to read variables
from stack frames, nor can it provide information about the source code
files and line numbers associated with program elements. But it can
provide the necessary information to implement a Type Finder. Assuming
symbol information is available (e.g. from kallsyms), then an Object
Finder can be implemented, which looks up names in the symbol table and
then uses libctf to find the type of the object.

This commit introduces a CTF type & object finder according to this
approach. To use it, a helper function is added, which allows a user to
provide the location of the vmlinux.ctfa file.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Reviewed-by: Junxiao Bi <junxiao.bi@oracle.com>