ctf: improve semantics of type & object lookup
Drgn provides a "filename" parameter to handle handle duplicate types,
by specifying the filename in which a type was defined. In general, CTF
cannot satisfy this API, especially for kernel CTF.

Kernel CTF has used this to mean "kernel module", but in order to avoid
errors when a filename is passed in, we have ignored errors. This means
we can't return errors when a module name is incorrect.

Add a trick to handle this: check if the string looks like a filename.
If so, for kernel CTF, ignore it. Otherwise, use it as the dictionary
name. With this change, we can then return errors when the specified
dictionary name is not present.

As part of this, to make the lookup more useful, we need to allow name
lookup to search the parent dicitonary.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>