ctf: fix up bitfield handling
Existing bitfield handling was broken in some rare cases, where a
bitfield was present but had a member at offset 0 and size which was a
multiple of 8. This fooled the CTF finder into believing it was not in a
bitfield, and later, the "Integer with bitfield info outside compound
type" error occurred. This is because dwarf2ctf encodes bitfield members
using the smallest byte size that can hold the bitfield member. Thus, we
falsely believed that some members were not bitfield members.

The fix is somewhat frustrating, but necessary: we must lookup the
canonical integer type by name, and check *its* byte size, to determine
whether a member is a bit field.

In the process, I cleaned up the handling of bitfields by a lot. I hope
that the resulting code is clearer.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>