ctf: add preliminary userspace support with libbfd
The ctf_open() API is only available with linked to libctf and libbfd.
When linked to libctf-nobfd, there is no ctf_open(), and we must instead
use ctf_arc_bufopen(), which requires that we have manually found the
CTF data section and symbol section.

For testing, it can be useful to build CTF-enabled userspace
applications and debug their core dumps with drgn. While userspace CTF
support isn't perfect, it works in simple cases. However, it requires
using ctf_open(), because userspace CTF is within an ELF file, not a
separate archive.

Thus, let's add a symbol "WITH_BFD" that switches drgn to using
ctf_open(). For now, this isn't actually hooked up to autotools. For
development, it allows me to add a "-DWITH_BFD" and enable the
function manually.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>