tests: add CTF testing
The CTF format needs to be tested, but it is primarily used by the
kernel, and it is currently out-of-tree in Oracle kernels (but being
upstreamed). Rather than adding several out-of-tree patches to the
vmtest subsystem to test kernel CTF directly, a simpler approach is to
test userspace CTF.

The current drgn CTF implementation isn't generic enough to support
userspace properly. However, it is good enough to load a simple core
dump with CTF data, without a symbol table. This gives us everything we
need to begin testing type representations.

Add a "test.c" file which can be built either with DWARF, GCC's CTF, or
dwarf2ctf. The two CTF implementations have subtle differences which
drgn supports, and thus need to each be tested.

Along with this test.c file, add some basic tests. The goal is that
these tests should pass on both DWARF and CTF. This is not even close to
a complete test suite, but it does incorporate several areas where the
CTF implementation needs careful attention: structs with bitfields, and
multidimensional arrays.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>