libdrgn: ctf: Handle buggy CTF multidimensional arrays
The CTF specification does not mention the order of indices for
multidimensional arrays, but that's most likely because it should be
self evident. For example:

    int array[3][4];

This is an array of 3 elements. Each element is an array of four
integers. Thus, the expected way that this should be represented in CTF
is:

    type X, CTF_K_ARRAY, ctr_contents=Y, ctr_nelems=3
    type Y, CTF_K_ARRAY, ctr_contents=Z, ctr_nelems=4
    type Z, CTF_K_INTEGER

However, CTF does not currently do things in this way. For CTF generated
by dwarf2ctf, multidimensional arrays are simply not supported (they are
emitted as single-dimension arrays, and the size is taken from the last
dimension of the array). This is a bug, but it will likely go unfixed.

For CTF generated by GCC, multidimensional arrays are emitted with
indices reversed. That is, they get emitted like this:

    type X, CTF_K_ARRAY, ctr_contents=Y, ctr_nelems=4
    type Y, CTF_K_ARRAY, ctr_contents=Z, ctr_nelems=3
    type Z, CTF_K_INTEGER

This, too, is a bug, and a very bad one! This bug will likely be fixed
soon, but due to the amount of already published CTF data, we will need
to support the broken ordering too. The fix will be accompanied by a
flag which clients can check to implement the correct behavior.

So, to handle this, let's reverse the indices on multidimensional
arrays. To do it properly, we have to eliminate the previous
mutual-recursion which was used to create multidimensional arrays, and
instead iteratively construct them.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>