ctf: fix slices used as canonical type names
When testing, I encountered a case where:

    prog.type("enum foo")

Gave the expected enumeration type, but where:

    prog.constant("FOO_VALUE")  # FOO_VALUE is a value of enum foo

Returned an object of type "enum <anonymous>". It turns out that at load
time, we had marked a sliced enum as the canonical version of the enum.
The sliced version had an absent name.

It's never correct to treat a sliced type as canonical: they exist
solely to re-use an existing type, but update its encoding. As a result,
we should update the startup sequence to avoid marking sliced types as
canonical. The best way to detect them is to test whether
ctf_type_reference() returns a non-error value: that indicates a slice,
except for typedefs, which need a more complex detection.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>