ctf: Add support for out-of-tree modules
Out-of-tree modules are well-supported by the CTF kernel using the
standard module building process. A ".ctf" section is included in the
ELF section for the loaded module, much like the relevant DWARF sections
in the debuginfo file. In the future, integration with the drgn module
API will make this much better integrated with all of drgn. But for now,
we rely on the fact that libctf with libbfd can find read the ELF file
and find the relevant data. Just like loading a standard CTF archive,
this is exposed as a private helper function from the _drgn module.

Supporting out-of-tree modules is critical for enabling kernel CTF tests
with the vmtest infrastructure, since many of drgn's tests use an
out-of-tree kernel module to create control data which can then be
tested against.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Reviewed-by: Junxiao Bi <junxiao.bi@oracle.com>