vmtest: add CTF support
While CTF is not yet part of the upstream kernel, Oracle regularly
publishes patches on each major release, which correspond to all changes
necessary to enable CTF (and thus, DTrace) on Linux. We can apply these
patches to the vmtest kernel in order to enable CTF debuginfo. Update
the vmtest.manage script to do just this.

To run tests with CTF support, we add options to the setup.py runner,
and we use these options to control the LinuxKernelTestCase. Most drgn
tests run without modification. However, some changes and skips are
necessary in areas which CTF cannot support:

1. Skip tests which read local variables & registers from a stack trace,
as this requires DWARF CFI.

2. Make changes so that tests pass when StackFrame.name is None, but
StackFrame.symbol() returns the relevant function symbol. ORC and
function pointer unwinds result in such stack frames.

3. Skip the vmcore tests which unwind the crashed CPU. Unfortunately, it
appears ORC for the relevant instruction has opcode UNKNOWN, so
__crash_kexec() cannot be unwound.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Reviewed-by: Junxiao Bi <junxiao.bi@oracle.com>