libdrgn, python: add SymbolIndex
The Symbol Finder API gives us the ability to register a dynamic
callback for symbol lookup. However, many common use cases are satisfied
by a simple static list of symbols. Correct and efficient lookup in this
simple case is rather tricky. Implement a new type, SymbolIndex, which
can take a list of symbols and index them for efficient lookup by name
or address.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Reviewed-by: Imran Khan <imran.f.khan@oracle.com>