libdrgn: add BTF type & object finder
Most kernels contain built-in BPF Type Information (BTF) describing all
kernel types. While this is intended for use by tracers and BPF
programs, it can also be used to debug a kernel without DWARF debug
info. Add support for BTF using libbpf, and support loading BTF from a
buffer or from the .BTF section.

Two new "btf" finders are added: one for types, and one for objects. The
BTF type finder is self-explanatory: it simply looks up types by name.
The object finder is less obvious: it looks up enum values (i.e.
constants), function objects (BTF_KIND_FUNC), and variables. The object
finder can only find variables for which a BTF_KIND_VAR and
BTF_KIND_DATASEC entry exist. Currently, this is only true for per-CPU
varables, although support exists in pahole to generate this data for
all variables.

Due to the limitations of variable availability, this support is not yet
useful for a lot of cases. Some basic functionality is not yet
available unless you create a "hardcoded" object finder, which combines
a well-known type declaration with the kallsyms symbol address for a
variable. An example of this is provided in drgn.helpers.linux.btf,
which provides enough information so that load_btf() can process the
module list and fully initialize drgn's module system. With this, it's
possible to perform stack traces and enumerate modules!

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>