cli: add CTF support
Running drgn scripts via "drgn -c TARGET -s DEBUGINFO script.py" is a
really handy part of drgn. As it is, CTF support can't be used in this
way, because drgn's CLI doesn't support CTF and it's not automatically
used.

We don't want to automatically use CTF, so let's instead add a new
argument, "-C". When provided, the "-s" argument is assumed to be a
"vmlinux.ctfa" file, and if "-s" is not provided, we just use the
default (which is usually the right thing to do anyway).

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Reviewed-by: Imran Khan <imran.f.khan@oracle.com>