libdrgn: util: add qsort_arg() helper function
In commit c8ff8728 ("Support systems without qsort_r"), usage of qsort_r
was eliminated because it is a glibc extension. There was discussion of
creating a utility function that implements qsort_r(), but the approach
described uses thread local variables, so it is not actually reentrant,
and it was dropped to avoid confusion.

However, upcoming commits will also prefer a comparator function which
takes an argument, and they also won't require a reentrant
implementation. Add this helper in with a name that shouldn't spark
confusion: qsort_arg().

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>
Reviewed-by: Junxiao Bi <junxiao.bi@oracle.com>