ctf: fix leaky enum constant creation
When creating enum constants, we should cache the enum type just like we
do for the normal type finding. The drgn_enum_from_ctf() function does
not do this, and so each time a user looks up a enum constant directly,
they get a new copy of the enum type. This is not a leak per-se, in that
drgn will clean up the types. But it is inefficient and could balloon
memory usage. Switch to the more general drgn_type_from_ctf_id() which
will performs caching.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>