orc_info: enable the use of built-in ORC
Since drgn typically assumes the presence of debug files, ORC has always
been loaded from the ELF file, rather than the program. However, ORC is
present in kernel core dumps, so it can be used when the debug file is
unavailable. Implement the ability to load built-in ORC for vmlinux and
kernel modules. We still prefer to load ORC from the debug file wherever
possible, because this is almost certainly faster.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>