elf_symtab: add support for .gnu_debugdata
The .gnu_debugdata section, also known as "MiniDebuginfo"[1], is an ELF
section which contains a second, compressed ELF file. This contained
file typically has all of its data stripped, except for a symbol table.
The symbols contained in this section are meant to be used in addition
to symbols that might be found in the ".dynsym" section.

For distributions like Fedora, this is intended to be used in
combination with .eh_frame information that is provided in the loadable
ELF file, so that stack traces can be automatically created for
userspace crashes, even without debuginfo. To that end, GDB has support
for these sections.

It makes sense for drgn to also include support for these symbol tables,
as we move toward supporting use cases where full DWARF data may not be
available.

Closes #465.

[1]: https://fedoraproject.org/wiki/Features/MiniDebugInfo

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>