libdrgn: aarch64: add linux fallback unwinding heuristics
The aarch64 frame pointer unwinder is able to unwind through exceptions
prior to v6.13, but it cannot recover the registers on the stack, and
thus it cannot mark stack frames as "interrupted".

Since 6.13, it is unable to unwind through exceptions at all, because of
a new signal stack frame which signals the presence of a pt_regs.

In the latter case, drgn can identify the pt_regs with confidence,
returning a stack frame with all registers identified, and the
"interrupted" flag set. In the former case, we can apply a heuristic to
detect whether the stack frame we see actually sits within one of the
"__irq_regs" percpu values.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>