libdrgn: python: fix race on nested drgn_begin_blocking()
If drgn_begin_blocking() is called without the GIL held (because we're
in a nested blocking section), and another thread has acquired the GIL,
then PyThreadState_GetUnchecked() can return the PyThreadState for that
thread, which we then mistakenly release. This causes all sorts of
crashes. I saw it spuriously when running setup.py test -K, which
downloads kernels in a separate thread while running local tests. I was
also able to reproduce it with the following:

  import threading
  import time

  from drgn import Program
  from tests.resources import get_resource

  def lol():
      while True:
          time.sleep(0.0001)  # Releases GIL.

  threading.Thread(target=lol, daemon=True).start()

  crashme_core = get_resource("crashme.core")
  crashme_path = get_resource("crashme")
  while True:
      prog = Program()
      prog.set_core_dump(crashme_core)
      prog.set_enabled_debug_info_finders([])
      prog.load_debug_info([crashme_path], main=True)

Note since CPython commit f8abfa331421 ("gh-103323: Get the "Current"
Thread State from a Thread-Local Variable (gh-103324)") (in v3.12),
PyThreadState_GetUnchecked() will never return the PyThreadState of
another thread, so this only affects earlier Python versions.

The documentation states that the GIL must be held when calling
PyThreadState_Get(). It is unclear if this also holds for
PyThreadState_GetUnchecked(). For the existing versions of CPython, it
looks like it will work without the GIL, but it's sketchy to rely on
this. Unfortunately, with the existing APIs, it's the best we can do.

Fixes: 48b962969540 ("libdrgn: python: allow nesting blocking callbacks")
Signed-off-by: Omar Sandoval <osandov@osandov.com>
(cherry picked from commit 2145ab8c03bfe52b22acd892547e96502db4103a)
Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>