libdrgn: linux_kernel: add drgn_program_is_irq_regs()
The kernel maintains a per-cpu variable "__irq_regs" which is set to the
struct pt_regs from the interrupted context on IRQ entry, and cleared on
exit. This can be used by unwinders to more reliably unwind through IRQ
frames, properly setting the "interrupted" flag and registers.

Add a new function which checks to see whether a pointer value
corresponds to the __irq_regs variable on any CPU. Since unwinding IRQ
stacks can never be reliable on live systems, this function only works
in non-live cases. We cache a list of non-NULL __irq_regs values on the
first call.

In order to do this, we need to access the cpu_present_mask. While drgn
has Python helpers for this, implementing them in C doesn't seem like a
significant improvement. For now, we just implement a basic helper
function that loads the cpu_present_mask, duplicating the Python logic.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>