libdrgn: dwarf_info: fix drgn_dwarf_die_iterator exhaustion case
If a drgn_dwarf_die_iterator is exhausted with subtree == 0, then the
vector of DIEs should be empty afterwards. Instead, we're leaving the
last CU DIE in the vector.

If the module has DWARF sections but no units (which doesn't happen in
practice but could be crafted), then we're leaving an uninitalized entry
in the vector, which could cause a crash when it is accessed.

Otherwise, this doesn't cause any user-visible problems:
drgn_dwarf_die_iterator_next is only used by
drgn_module_find_dwarf_scopes(), which is only used for stack traces.
The CU DIE is only used for drgn_stack_frame_source() and
drgn_stack_frame_find_object(). In the former, the CU DIE won't contain
the program counter, so it fails as expected. In the latter, the CU DIE
is ignored because there is no function scope. However, an upcoming API
for mapping addresses to source code locations is sensitive to bogus
CUs.

Fixes: b6d810b344e3 ("libdrgn: debug_info: add DWARF DIE iterator")
Signed-off-by: Omar Sandoval <osandov@osandov.com>
(cherry picked from commit 0ed1a80b91e9bdc4901d1f48156a520672f49760)
Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>